fix(axswift): add getSelectedTextFrame() and update AXSwift dependency
Introduces a new `getSelectedTextFrame()` function to retrieve the screen coordinates (NSRect) of the currently selected text using the Accessibility API.

To enable this functionality, this commit updates the `AXSwift` dependency to point to the `clavierorg/AXSwift` fork (v0.3.4). This version includes a critical fix for `parameterizedAttribute` which allows `.boundsForRangeParameterized` to work correctly.

This resolves the issue described in: https://github.com/tmandry/AXSwift/issues/44